+++ dokidokimd/consts.py
@@ -1,0 +1,7 @@
+from PyQt5.QtGui import QColor
+
+QCOLOR_DOWNLOADED = QColor(23, 150, 200, 250)
+QCOLOR_CONVERTERR = QColor(11, 120, 10, 250)
+QCOLOR_HIGHLIGHT = QColor(142, 45, 197, 255)
+QCOLOR_WHITE = QColor(255, 255, 255, 220)
+QCOLOR_DARK = QColor(33, 33, 33, 220)
