+++ dokidokimd/main.py
@@ -1,5 +1,5 @@
+from pyqt.gui import start_gui
 from tools.kz_logger import init_logger, PROJECT_NAME
-from gui import start_gui
 
 init_logger()
 start_gui(PROJECT_NAME)
