+++ setup.py
@@ -7,7 +7,7 @@ from setuptools import setup
 name = 'dokidokimd'
 
 DIR = join(dirname(abspath(__file__)), name)
-VERSION = '0.1.0'
+VERSION = '0.2.0'
 
 with open(join(dirname(abspath(__file__)), 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
@@ -27,10 +27,10 @@ setup(
     description='Python manga downloader and PDF creator.',
     long_description=long_description,
     license='BSD 3-Clause',
-    keywords='tui urwid python-manga-downloader python python3 manga manga-downloader',
+    keywords='python-manga-downloader python python3 manga manga-downloader',
     author='Konrad Ziarko',
     author_email='konrad.ziarko@protonmail.ch',
-    url='https://github.com/Konrad-Ziarko/PyPractice',
+    url='https://github.com/Konrad-Ziarko/DokiDokiMD',
     download_url='https://github.com/Konrad-Ziarko/DokiDokiMD/archive/v{0}.zip'.format(VERSION),
     packages=['dokidokimd'],
     package_data={'': package_files()},
