+++ tests/functional/test_create_pdf_from_jpegs/test_make_pdf.py
@@ -9,9 +9,7 @@ def test_make_pdf():
 
     pdf.add_dir(join(directory_name, 'images'))
 
-    pdf.make_pdf('test')
-
-    pdf.save_pdf(join(directory_name, 'results', 'test_result.pdf'))
+    pdf.make_pdf('test', join(directory_name, 'results', 'test_result.pdf'))
 
     assert isfile(join(directory_name, 'results', 'test_result.pdf')) is True
     assert getsize(join(directory_name, 'results', 'test_result.pdf')) > 0
